+++ Bugsnag/Assets/Bugsnag/Runtime/Native/Cocoa/NativeClient.cs
@@ -24,7 +24,11 @@ namespace BugsnagUnity
         private bool _registeredForSessionCallbacks;
 
         private LoadedImages loadedImages = new LoadedImages();
-
+        private const string DsymUuidsKey = "dsymUUIDs";
+        private const string IsLaunchingKey = "isLaunching";
+        private const string JailbrokenKey = "jailbroken";
+        private const string OsBuildKey = "osBuild";
+        
         public NativeClient(Configuration configuration)
         {
             _instance = this;
@@ -259,12 +263,20 @@ namespace BugsnagUnity
             var dictionary = ((JsonObject)SimpleJson.DeserializeObject(result)).GetDictionary();
             foreach (var pair in dictionary)
             {
-                if (pair.Key == "isLaunching")
+                if (pair.Key == IsLaunchingKey)
                 {
                     if (pair.Value != null)
                     {
                         var stringValue = (pair.Value as string).ToLower();
                         app.Add(pair.Key, stringValue == "true");
+                    }
+                }
+                else if (pair.Key == DsymUuidsKey)
+                {
+                    var uuids = pair.Value as JsonArray;
+                    if (uuids != null && uuids.Count > 0)
+                    {
+                        app.DsymUuid = uuids[0] as string;
                     }
                 }
                 else
@@ -285,7 +297,7 @@ namespace BugsnagUnity
             var dictionary = ((JsonObject)SimpleJson.DeserializeObject(result)).GetDictionary();
             foreach (var pair in dictionary)
             {
-                if (pair.Key == "jailbroken")
+                if (pair.Key == JailbrokenKey)
                 {
                     if (pair.Value != null)
                     {
@@ -293,7 +305,7 @@ namespace BugsnagUnity
                         device.Add(pair.Key, stringValue == "true");
                     }
                 }
-                else if (pair.Key == "osBuild")
+                else if (pair.Key == OsBuildKey)
                 {
                     device.RuntimeVersions.AddToPayload(pair.Key, pair.Value);
                 }