+++ features/steps/unity_steps.rb
@@ -234,7 +234,7 @@ Then("all possible parameters have been edited in a callback") do
     And the event "app.bundleVersion" equals "BundleVersion"
     And the event "app.binaryArch" equals "BinaryArch"
     And the event "app.codeBundleId" equals "CodeBundleId"
-    And the event "app.dsymUuid" equals "DsymUuid"
+    And the event "app.dsymUUIDs.0" equals "DsymUuid"
     And the event "app.inForeground" is false
     And the event "app.isLaunching" is false