+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 - Fix Android context being overwritten by Activity
 [#977](https://github.com/bugsnag/bugsnag-unity/pull/977)
 
+- Added `DsymUUID` in managed event app metadata
+[#997](https://github.com/bugsnag/bugsnag-unity/pull/997)
+
 ## 8.9.1 (2026-04-01)
 
 ### Bug Fixes